moved the google places autocomplete logic input into its own component, also added a logo temp image into the project
